Enough refactoring
