+++ src/player/player_spawner.rs
@@ -4,8 +4,8 @@ use bevy::input::gamepad::GamepadButton;
 use bevy::prelude::*;
 use bevy_xpbd_2d::prelude::*;
 
+use crate::global::AimType;
 use crate::global::*;
-use crate::{global::AimType, projectile::projectile_spawner::*};
 
 use super::{AttackTimer, Player, ATTACK_SPEED, PLAYER_SIZE, PLAYER_SPEED};
 
@@ -34,27 +34,6 @@ pub fn spawn_player(mut commands: Commands) {
             aim_type: AimType::Random,
         },
     ));
-}
-
-pub fn player_attack(
-    time: Res<Time>,
-    mut attack_timer_query: Query<&mut AttackTimer>,
-    mut player_counter: PlayerProjectileSpawner,
-) {
-    let mut attack_timer = attack_timer_query.get_single_mut().unwrap();
-    attack_timer.timer.tick(time.delta());
-    if attack_timer.timer.finished() {
-        player_counter.spawn_projectile();
-
-        //let mut rng = rand::thread_rng();
-        /*let sound_effect;
-        // Randomly play one of the two sound effects.
-        if rng.gen_bool(0.5) {
-            sound_effect = "audio/tap.ogg";
-        } else {
-            sound_effect = "audio/ti.ogg";
-        }*/
-    }
 }
 
 // TODO: leafwing-input-manager
